+++ src/ethereum/EthereumEngine.ts
@@ -119,9 +119,10 @@ export class EthereumEngine extends CurrencyEngine<
     this.ethNetwork = new EthereumNetwork(this)
     if (this.networkInfo.optimismRollup === true) {
       this.optimismRollupParams = {
-        gasPriceL1Wei: '1000000000',
-        fixedOverhead: '2100',
-        dynamicOverhead: '1000000'
+        baseFee: '1000000000',
+        baseFeeScalar: '1368',
+        blobBaseFee: '1',
+        blobBaseFeeScalar: '659851'
       }
     }
     this.networkFees = this.networkInfo.defaultNetworkFees
@@ -551,7 +552,7 @@ export class EthereumEngine extends CurrencyEngine<
 
     const oracleContractAddress = '0x420000000000000000000000000000000000000F'
 
-    // L1GasPrice
+    // Base fee
     try {
       const params = {
         to: oracleContractAddress,
@@ -565,7 +566,7 @@ export class EthereumEngine extends CurrencyEngine<
 
       this.optimismRollupParams = {
         ...this.optimismRollupParams,
-        gasPriceL1Wei: ceil(
+        baseFee: ceil(
           mul(
             hexToDecimal(result.result),
             '1.25' // maxGasPriceL1Multiplier
@@ -574,27 +575,67 @@ export class EthereumEngine extends CurrencyEngine<
         )
       }
     } catch (e: any) {
-      this.log.warn('Failed to update l1GasPrice', e)
+      this.log.warn('Failed to update base fee', e)
     }
 
-    // Dynamic overhead (scalar)
+    // Blob base fee
     try {
       const params = {
         to: oracleContractAddress,
-        data: '0xf45e65d8' // dynamic overhead method
+        data: '0xf8206140' // L1 Blob base fee method
       }
       const response = await this.ethNetwork.multicastRpc('eth_call', [
         params,
         'latest'
       ])
-
       const result = asRpcResultString(response.result)
+
       this.optimismRollupParams = {
         ...this.optimismRollupParams,
-        dynamicOverhead: hexToDecimal(result.result)
+        blobBaseFee: ceil(hexToDecimal(result.result), 0)
       }
     } catch (e: any) {
-      this.log.warn('Failed to update dynamicOverhead', e)
+      this.log.warn('Failed to update blob base fee', e)
+    }
+
+    // Base fee scalar
+    try {
+      const params = {
+        to: oracleContractAddress,
+        data: '0xc5985918' // base fee scalar method
+      }
+      const response = await this.ethNetwork.multicastRpc('eth_call', [
+        params,
+        'latest'
+      ])
+      const result = asRpcResultString(response.result)
+
+      this.optimismRollupParams = {
+        ...this.optimismRollupParams,
+        baseFeeScalar: hexToDecimal(result.result)
+      }
+    } catch (e: any) {
+      this.log.warn('Failed to update base fee scalar', e)
+    }
+
+    // Blob base fee scalar
+    try {
+      const params = {
+        to: oracleContractAddress,
+        data: '0x68d5dca6' // blob base fee method
+      }
+      const response = await this.ethNetwork.multicastRpc('eth_call', [
+        params,
+        'latest'
+      ])
+      const result = asRpcResultString(response.result)
+
+      this.optimismRollupParams = {
+        ...this.optimismRollupParams,
+        blobBaseFeeScalar: hexToDecimal(result.result)
+      }
+    } catch (e: any) {
+      this.log.warn('Failed to update blob base fee scalar', e)
     }
   }
 
@@ -813,14 +854,15 @@ export class EthereumEngine extends CurrencyEngine<
         const maxSpendableBeforeRollupFee = sub(balance, primaryNetworkFee)
 
         const txData: CalcOptimismRollupFeeParams = {
+          baseFee: this.optimismRollupParams.baseFee,
+          baseFeeScalar: this.optimismRollupParams.baseFeeScalar,
+          blobBaseFee: this.optimismRollupParams.blobBaseFee,
+          blobBaseFeeScalar: this.optimismRollupParams.blobBaseFeeScalar,
           nonce: this.otherData.unconfirmedNextNonce,
-          gasPriceL1Wei: this.optimismRollupParams.gasPriceL1Wei,
           gasLimit: miningFees.gasLimit,
           to: publicAddress,
           value: decimalToHex(maxSpendableBeforeRollupFee),
-          chainParams: this.networkInfo.chainParams,
-          dynamicOverhead: this.optimismRollupParams.dynamicOverhead,
-          fixedOverhead: this.optimismRollupParams.fixedOverhead
+          chainParams: this.networkInfo.chainParams
         }
         rollupFee = calcOptimismRollupFees(txData)
       } else if (this.networkInfo.arbitrumRollupParams != null) {
@@ -1042,15 +1084,16 @@ export class EthereumEngine extends CurrencyEngine<
     //  Optimism-style L1 fees are deducted automatically from the account. Arbitrum-style L1 gas must be included in the transaction object.
     if (this.optimismRollupParams != null) {
       const txData: CalcOptimismRollupFeeParams = {
+        baseFee: this.optimismRollupParams.baseFee,
+        baseFeeScalar: this.optimismRollupParams.baseFeeScalar,
+        blobBaseFee: this.optimismRollupParams.blobBaseFee,
+        blobBaseFeeScalar: this.optimismRollupParams.blobBaseFeeScalar,
         nonce: otherParams.nonceUsed,
-        gasPriceL1Wei: this.optimismRollupParams.gasPriceL1Wei,
         gasLimit: otherParams.gas,
         to: otherParams.to[0],
         value: value,
         data: otherParams.data,
-        chainParams: this.networkInfo.chainParams,
-        dynamicOverhead: this.optimismRollupParams.dynamicOverhead,
-        fixedOverhead: this.optimismRollupParams.fixedOverhead
+        chainParams: this.networkInfo.chainParams
       }
       l1Fee = calcOptimismRollupFees(txData)
     } else if (this.networkInfo.arbitrumRollupParams != null) {