+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- changed: Use EIP-4844 fee calculation for Optimism and Base
+
 ## 3.2.6 (2024-03-20)
 
 - changed: Update subscan transfers endpoint url